requirements.txt
streamlit>=1.36.0
faiss-gpu>=1.7.4
sentence-transformers>=3.0.0
openai>=1.30.0
python-dotenv>=1.0.1
numpy>=1.26.0
requests>=2.32.0

